+++ widgets/src/main/kotlin/com/commonsense/android/kotlin/views/databinding/activities/BaseDatabindingActivityWithData.kt
@@ -21,4 +21,5 @@ abstract class BaseDatabindingActivityWithData<out view : ViewDataBinding, out I
         binding.executePendingBindings()
         useBinding()
     }
+
 }